+++ src/index.ts
@@ -24,7 +24,6 @@ export {Uuid, UploadAPIInterface} from './api/types'
 export {FileUploadInterface} from './fileFrom/types'
 export {UploadClientInterface} from './UploadClient'
 export {RequestOptionsInterface, RequestInterface} from './api/request/types'
-export {FileFromEnum} from './fileFrom/types'
 
 /* Low-Level API */
 export {request}