+++ test/fileFrom/fileFromUploaded.spec.ts
@@ -1,6 +1,5 @@
 import * as factory from '../_fixtureFactory'
 import fileFrom from '../../src/fileFrom/fileFrom'
-import {FileFromEnum} from '../../src/fileFrom/types'
 import {getSettingsForTesting} from '../_helpers'
 
 describe('fileFrom', () => {
@@ -11,7 +10,7 @@ describe('fileFrom', () => {
       const settings = getSettingsForTesting({
         publicKey: factory.publicKey('image'),
       })
-      const file = await fileFrom(FileFromEnum.Uploaded, uuid, settings)
+      const file = await fileFrom(uuid, settings)
 
       expect(file.cdnUrl).toBeTruthy()
     })
@@ -20,7 +19,7 @@ describe('fileFrom', () => {
       const settings = getSettingsForTesting({
         publicKey: factory.publicKey('image'),
       })
-      const filePromise = fileFrom(FileFromEnum.Uploaded, uuid, settings)
+      const filePromise = fileFrom(uuid, settings)
 
       setTimeout(() => {
         filePromise.cancel()
@@ -37,7 +36,7 @@ describe('fileFrom', () => {
         doNotStore: true,
         fileName: 'newFileName.jpg',
       })
-      const file = await fileFrom(FileFromEnum.Uploaded, uuid, settings)
+      const file = await fileFrom(uuid, settings)
 
       expect(file.name).toEqual('newFileName.jpg')
     })
@@ -47,7 +46,7 @@ describe('fileFrom', () => {
         const settings = getSettingsForTesting({
           publicKey: factory.publicKey('image'),
         })
-        const filePromise = fileFrom(FileFromEnum.Uploaded, uuid, settings)
+        const filePromise = fileFrom(uuid, settings)
 
         setTimeout(() => {
           filePromise.cancel()
@@ -71,7 +70,7 @@ describe('fileFrom', () => {
         const settings = getSettingsForTesting({
           publicKey: factory.publicKey('image'),
         })
-        const filePromise = fileFrom(FileFromEnum.Uploaded, uuid, settings)
+        const filePromise = fileFrom(uuid, settings)
 
         filePromise.onProgress = (progress) => {
           const {value} = progress
@@ -88,7 +87,7 @@ describe('fileFrom', () => {
         const settings = getSettingsForTesting({
           publicKey: factory.publicKey('image'),
         })
-        const filePromise = fileFrom(FileFromEnum.Uploaded, uuid, settings)
+        const filePromise = fileFrom(uuid, settings)
 
         filePromise.onUploaded = () => {
           done()
@@ -102,7 +101,7 @@ describe('fileFrom', () => {
         const settings = getSettingsForTesting({
           publicKey: factory.publicKey('image'),
         })
-        const filePromise = fileFrom(FileFromEnum.Uploaded, uuid, settings)
+        const filePromise = fileFrom(uuid, settings)
 
         filePromise.onReady = () => {
           done()